feat: shimmer import progress message
